+++ api/src/main/java/io/strimzi/api/kafka/model/kafka/PersistentClaimStorage.java
@@ -24,7 +24,7 @@ import java.util.Map;
         editableEnabled = false,
         builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
-@JsonPropertyOrder({"id", "type", "size", "class", "selector", "deleteClaim", "overrides"})
+@JsonPropertyOrder({"id", "type", "size", "kraftMetadata", "class", "selector", "deleteClaim", "overrides"})
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
@@ -66,6 +66,21 @@ public class PersistentClaimStorage extends SingleVolumeStorage {
 
     public void setSize(String size) {
         this.size = size;
+    }
+
+    @Override
+    @Description("Specifies whether this volume should be used for storing KRaft metadata. " +
+            "This property is optional. " +
+            "When set, the only currently supported value is `shared`. " +
+            "At most one volume can have this property set.")
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    public KRaftMetadataStorage getKraftMetadata() {
+        return super.getKraftMetadata();
+    }
+
+    @Override
+    public void setKraftMetadata(KRaftMetadataStorage kraftMetadata) {
+        super.setKraftMetadata(kraftMetadata);
     }
 
     @JsonProperty("class")